requirements.txt
python-telegram-bot==22.0
watchdog
gspread
oauth2client


